[IMP] web: convert mobile datetime_input date test to hoot
task-4028335
